Untested attempt at adding a group controller step before choosing a revonarch.
